+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `hideApprovalDenyButton` to `approvalStepNodeProps`
+- `EXCEL_ADD_ROW` form workflow event to validation
 
 ## [12.3.1] - 2025-11-07